+++ src/main/java/com/pyding/deathlyhallows/events/DHEvents.java
@@ -22,9 +22,9 @@ import com.pyding.deathlyhallows.entities.EntityAbsoluteDeath;
 import com.pyding.deathlyhallows.entities.EntityNimbus;
 import com.pyding.deathlyhallows.integrations.DHIntegration;
 import com.pyding.deathlyhallows.items.DHItems;
-import com.pyding.deathlyhallows.items.ItemBaubleResurrectionStone;
 import com.pyding.deathlyhallows.items.ItemDeadlyPrism;
 import com.pyding.deathlyhallows.items.ItemNimbus;
+import com.pyding.deathlyhallows.items.baubles.ItemBaubleResurrectionStone;
 import com.pyding.deathlyhallows.network.DHPacketProcessor;
 import com.pyding.deathlyhallows.network.packets.PacketAnimaMobRender;
 import com.pyding.deathlyhallows.network.packets.PacketPlayerRender;
