+++ src/main/java/com/pyding/deathlyhallows/items/ItemTrickOrTreat.java
@@ -7,6 +7,7 @@ import com.emoniph.witchery.item.ItemVanillaPotion;
 import com.emoniph.witchery.util.ParticleEffect;
 import com.emoniph.witchery.util.SoundEffect;
 import com.pyding.deathlyhallows.integrations.DHIntegration;
+import com.pyding.deathlyhallows.items.baubles.ItemBaubleResurrectionStone;
 import com.pyding.deathlyhallows.utils.DHUtils;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
